# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Inbox Shore API
  description: API documentation for Inbox Shore application
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/AleHundred/inbox-shore-git
servers:
  - url: http://localhost:3001/api
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Ticket:
      type: object
      properties:
        id:
          type: string
          example: ticket123
        subject:
          type: string
          example: Need help with login
        status:
          type: string
          example: open
        customerEmail:
          type: string
          format: email
          example: customer@example.com
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /requests:
    get:
      tags: [Tickets]
      summary: Get all tickets
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tickets
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  tickets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ticket'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /timeline:
    get:
      tags: [Tickets]
      summary: Get ticket timeline/details
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: ticketId
          required: true
          schema:
            type: string
          description: ID of the ticket to get details for
      responses:
        '200':
          description: Ticket timeline data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          description: Missing ticketId parameter
        '404':
          description: Ticket not found
...